Since playlist_t is using size_t, change two unsigned longs to size_t as well.
git-svn-id: https://svn.xiph.org/trunk/ezstream@13467 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ src/playlist.c
@@ -82,7 +82,7 @@ playlist_add(playlist_t *pl, const char *entry)
 	}
 
 	if (pl->size / sizeof(char *) <= num) {
-		unsigned long	i;
+		size_t	i;
 
 		pl->list = xrealloc(pl->list, 2, pl->size);
 		pl->size = 2 * pl->size;
@@ -438,8 +438,8 @@ playlist_reread(playlist_t **plist)
 void
 playlist_shuffle(playlist_t *pl)
 {
-	unsigned long	 d, i, range;
-	char		*temp;
+	size_t	 d, i, range;
+	char	*temp;
 
 	if (pl == NULL) {
 		printf("%s: playlist_shuffle(): Internal error: NULL argument\n",
